Refactor world streaming into shared modules
